+++ NEWS.md
@@ -1,5 +1,12 @@
-CohortMethod 5.1.1
+CohortMethod 5.2.0
 ==================
+
+Changes:
+
+1. The `computeCovariateBalance()` function now also computes standardized difference of mean comparing cohorts before and after PS adjustment, which can inform on generalizability.
+
+2. Improved computation of overall standard deviation when computing covariate balance. Should produce more accurate balance estimations.
+
 
 Bugfixes: